add xml and json I/O

+++ src/main/java/com/koror/app/repository/GroupRepository.java
@@ -1,5 +1,7 @@
 package com.koror.app.repository;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import com.koror.app.api.repository.IGroupRepository;
 import com.koror.app.entity.Group;
 
@@ -12,6 +14,12 @@ import java.util.Map;
 public class GroupRepository implements IGroupRepository {
 
     private final Map<String, Group> groupMap = new HashMap<>();
+
+    private final String pathXml = "data/group/data_group.xml";
+
+    private final String pathJson = "data/group/data_group.json";
+
+    private final String pathBin = "data/group/data_group.tmp";
 
     @Override
     public void addGroup(final Group group) {
@@ -43,25 +51,55 @@ public class GroupRepository implements IGroupRepository {
     }
 
     @Override
-    public void saveData() {
-        try (final ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream("data_group.tmp"))) {
-            final File f = new File("data_group.tmp");
-            if (f.isFile())
-                f.delete();
-            oos.writeObject(getGroupList());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public void saveDataSerialization() throws IOException {
+        final FileOutputStream fos = new FileOutputStream(pathBin);
+        final ObjectOutputStream oos = new ObjectOutputStream(fos);
+        final File f = new File(pathBin);
+        new File("data/group/").mkdirs();
+        if (f.isFile()) f.delete();
+        oos.writeObject(getGroupList());
     }
 
     @Override
-    public void loadData() {
-        try (final ObjectInputStream ois = new ObjectInputStream(new FileInputStream("data_group.tmp"))) {
-            final List<Group> tasks = (List<Group>) ois.readObject();
-            for (Group group : tasks)
-                groupMap.put(group.getId(), group);
-        } catch (IOException | ClassNotFoundException e) {
-            e.printStackTrace();
-        }
+    public void loadDataSerialization() throws IOException, ClassNotFoundException{
+        final FileInputStream fis = new FileInputStream(pathBin);
+        final ObjectInputStream ois = new ObjectInputStream(fis);
+        final List<Group> tasks = (List<Group>) ois.readObject();
+        for (Group group : tasks) groupMap.put(group.getId(), group);
     }
+
+    @Override
+    public void saveDataXml() throws IOException {
+        final File f = new File(pathXml);
+        new File("data/group/").mkdirs();
+        if (f.isFile()) f.delete();
+        final ObjectMapper objectMapper = new XmlMapper();
+        final Group[] listGroup = getGroupList().toArray(new Group[getGroupList().size()]);
+        objectMapper.writeValue(new File(pathXml), listGroup);
+    }
+
+    @Override
+    public void loadDataXml() throws IOException {
+        final ObjectMapper objectMapper = new XmlMapper();
+        final Group[] listGroup = objectMapper.readValue(new File(pathXml), Group[].class);
+        for (Group group : listGroup) addGroup(group);
+    }
+
+    @Override
+    public void saveDataJson() throws IOException {
+        final File f = new File(pathJson);
+        new File("data/group/").mkdirs();
+        if (f.isFile()) f.delete();
+        final ObjectMapper objectMapper = new ObjectMapper();
+        final Group[] listGroup = getGroupList().toArray(new Group[getGroupList().size()]);
+        objectMapper.writeValue(new File(pathJson), listGroup);
+    }
+
+    @Override
+    public void loadDataJson() throws IOException{
+        final ObjectMapper objectMapper = new ObjectMapper();
+        final Group[] listGroup = objectMapper.readValue(new File(pathJson), Group[].class);
+        for (Group task : listGroup) addGroup(task);
+    }
+
 }
